refactor(modal): rename context file
